@@ -1,19 +1,54 @@
-UnifiedWarehouse::Application.configure do
+Rails.application.configure do
+  # Settings specified here will take precedence over those in config/application.rb.
+
+  # The test environment is used exclusively to run your application's
+  # test suite. You never need to work with it otherwise. Remember that
+  # your test database is "scratch space" for the test suite and is wiped
+  # and recreated between test runs. Don't rely on the data there!
+
   # NOTE: Much of this can probably be removed as we're not really using Rails, more AR.
   config.cache_classes = true
-  # config.whiny_nils = true
-  config.consider_all_requests_local       = true
-  config.action_controller.perform_caching = false
-  config.action_mailer.raise_delivery_errors = false
-  config.active_support.deprecation = :log
-  # config.action_dispatch.best_standards_support = :builtin
-  # config.assets.compress = false
-  config.assets.debug = true
+
+  # Do not eager load code on boot. This avoids loading your whole application
+  # just for the purpose of running a single test. If you are using a tool that
+  # preloads Rails for running tests, you may have to set it to true.
+  config.eager_load = false
+
+  # Configure public file server for tests with Cache-Control for performance.
+  config.public_file_server.enabled = true
+  config.public_file_server.headers = {
+    'Cache-Control' => "public, max-age=#{1.hour.to_i}"
+  }
+
+  # Show full error reports and disable caching.
+  # config.consider_all_requests_local       = true
+  # config.action_controller.perform_caching = false
+
+  # Raise exceptions instead of rendering exception templates.
+  # config.action_dispatch.show_exceptions = false
+
+  # Disable request forgery protection in test environment.
+  # config.action_controller.allow_forgery_protection = false
+
+  # Store uploaded files on the local file system in a temporary directory
+  # config.active_storage.service = :test
+
+  # config.action_mailer.raise_delivery_errors = false
+
+  # config.assets.debug = true
   config.log_level = :warn
 
-  # Here is some ActiveRecord configuration that is useful
-  # config.active_record.mass_assignment_sanitizer         = :strict
-  # config.active_record.auto_explain_threshold_in_seconds = 0.5
+  # Tell Action Mailer not to deliver emails to the real world.
+  # The :test delivery method accumulates sent emails in the
+  # ActionMailer::Base.deliveries array.
+
+  # config.action_mailer.delivery_method = :test
+
+  # Print deprecation notices to the stderr.
+  config.active_support.deprecation = :stderr
+
+  # Raises error for missing translations
+  # config.action_view.raise_on_missing_translations = true
 
   # Configure the numeric tolerance
   config.numeric_tolerance = 0.05
