@@ -1,10 +1,24 @@
+# Use to automatically test associations
+# @example A simple resource
+# ```
+# it_behaves_like 'belongs to', [
+#       :sample # It belongs to as sample
+#     ], nil
+# ```
+# @example A nested resource
+# ```
+# it_behaves_like 'belongs to', [
+#       :sample # It belongs to as sample
+#     ], { lanes: :controls } # The Json keys to follow to get to the relationship
+# ```
 shared_examples_for 'belongs to' do |belonging_owners, belonging_owned|
   let!(:json_handler) { described_class.send(:json) }
 
   def find_target(json, graph)
     return json if graph.nil?
     return json[graph.to_s].first if graph.is_a?(Symbol)
-    return json[graph.first.to_s].first if (graph.count == 1 && graph.first.is_a?(Symbol))
+    return json[graph.first.to_s].first if graph.count == 1 && graph.first.is_a?(Symbol)
+
     k, v = graph.first
     find_target(json[k.to_s].first, v)
   end
@@ -12,7 +26,8 @@ shared_examples_for 'belongs to' do |belonging_owners, belonging_owned|
   def find_handler(handler, graph)
     return handler if graph.nil?
     return handler.nested_models[graph] if graph.is_a?(Symbol)
-    return handler.nested_models[graph.first] if (graph.count == 1 && graph.first.is_a?(Symbol))
+    return handler.nested_models[graph.first] if graph.count == 1 && graph.first.is_a?(Symbol)
+
     k, v = graph.first
     find_handler(handler.nested_models[k], v)
   end
@@ -24,7 +39,7 @@ shared_examples_for 'belongs to' do |belonging_owners, belonging_owned|
   let!(:owning_json) { find_target(json, belonging_owned) }
 
   belonging_owners.each do |owner|
-    context "#{owner}" do
+    context owner.to_s do
       let(:owning_handler) { find_handler(json_handler, belonging_owned) }
 
       let(:id)   { owning_json[:"#{owner}_id"] }
@@ -39,7 +54,7 @@ shared_examples_for 'belongs to' do |belonging_owners, belonging_owned|
   end
 
   after(:each) do
-    content = subject.is_a?(Array) ? subject.detect { |json| json.class == owning_handler } : subject
+    content = subject.is_a?(Array) ? subject.detect { |json| json.instance_of?(owning_handler) } : subject
     described_class.send(:create_or_update, content)
   end
 end
